# Add mod resource packs caching for faster reloads.