--- a/library/core/resource_loader/src/main/java/org/quiltmc/qsl/resource/loader/api/GroupResourcePack.java
+++ b/library/core/resource_loader/src/main/java/org/quiltmc/qsl/resource/loader/api/GroupResourcePack.java
@@ -184,7 +184,7 @@ public abstract class GroupResourcePack implements ResourcePack {
 		 * @param type         the resource type of this resource pack
 		 * @param basePack     the base resource pack
 		 * @param packs        the additional packs
-		 * @param basePriority {@code true} if the base resource pack has priority over the additional packs, otherwise {@code false}.
+		 * @param basePriority {@code true} if the base resource pack has priority over the additional packs, or {@code false} otherwise.
 		 *                     Ignored if the base resource pack is already present in the list
 		 */
 		public Wrapped(ResourceType type, ResourcePack basePack, List<ResourcePack> packs, boolean basePriority) {