Also use the java.io.File API when possible instead of NIO exists.

## Some timings for comparison sake:

### Dev Env

#### First Reload

| Method | Time |
|:---------:|-------:| 
| No Cache, no File API | 8372 ms |
| No Cache, File API | 5664 ms |
| Cache, no File API | 8252 ms |
| Cache, File API | 8195 ms |
| Tree Cache, no File API | 6263 ms |
| Tree Cache, File API | 6178 ms |

#### World Reload

| Method | Time |
|:---------:|-------:| 
| No Cache, no File API | 1409 ms |
| No Cache, File API | 1260 ms |
| Cache, no File API | 1304 ms |
| Cache, File API | 1487 ms |
| Tree Cache, no File API | 1275 ms |
| Tree Cache, File API | 1187 ms |

### Blanketcon modpack

#### First Reload

| Method | Time |
|:---------:|-------:| 
| No Cache | 27025 ms |
| Cache | 21641 ms |
| Tree Cache | 20378 ms |

#### World Reload

| Method | Time |
|:---------:|-------:| 
| No Cache | 24566ms |
| Cache | 20721ms - 21781ms |
| Tree Cache | 19679ms - 19917ms |

In the blanketcon modpack this manages to remove almost 7 seconds of startup time!

Cache corresponds to a much simpler cache implementation with a simple ConcurrentHashMap, Tree Cache is the cache implemented in this PR.

For the dev env, only the Minecraft resource pack itself is cached by this system since the QSL dev env only has this "mod" that is eligible for caching.
Mods in dev envs that may have mutable resources via debugging will not use the cache.